refactor pulling out feasibility from feasable intent service
