+++ src/intent/feasable-intent.service.ts
@@ -3,19 +3,12 @@ import { EcoConfigService } from '../eco-configs/eco-config.service'
 import { JobsOptions, Queue } from 'bullmq'
 import { InjectQueue } from '@nestjs/bullmq'
 import { QUEUES } from '../common/redis/constants'
-import { TransactionTargetData, UtilsIntentService } from './utils-intent.service'
-import { BalanceService } from '../balance/balance.service'
+import { UtilsIntentService } from './utils-intent.service'
 import { EcoLogMessage } from '../common/logging/eco-log-message'
-import { EcoError } from '../common/errors/eco-error'
-import { IntentSourceModel } from './schemas/intent-source.schema'
 import { getIntentJobId } from '../common/utils/strings'
-import { Solver } from '../eco-configs/eco-config.types'
-import { InjectModel } from '@nestjs/mongoose'
-import { Model } from 'mongoose'
 import { Hex } from 'viem'
-import { CallDataInterface, getERC20Selector, RewardTokensInterface } from '../contracts'
 import { QuoteIntentModel } from '@/quote/schemas/quote-intent.schema'
-import { ValidationIntentInterface } from '@/intent/validation.sevice'
+import { FeasibilityService } from '@/intent/feasibility.service'
 
 /**
  * Service class for getting configs for the app
@@ -24,20 +17,15 @@ import { ValidationIntentInterface } from '@/intent/validation.sevice'
 export class FeasableIntentService implements OnModuleInit {
   private logger = new Logger(FeasableIntentService.name)
   private intentJobConfig: JobsOptions
-  private FEE_BASE = 1000n
-  private fee: bigint
   constructor(
     @InjectQueue(QUEUES.SOURCE_INTENT.queue) private readonly intentQueue: Queue,
-    @InjectModel(IntentSourceModel.name) private intentModel: Model<IntentSourceModel>,
-    private readonly balanceService: BalanceService,
     private readonly utilsIntentService: UtilsIntentService,
+    private readonly feasableService: FeasibilityService,
     private readonly ecoConfigService: EcoConfigService,
   ) {}
 
   async onModuleInit() {
     this.intentJobConfig = this.ecoConfigService.getRedis().jobs.intentJobConfig
-    //todo get this from config or service per token/chain
-    this.fee = 1000n
   }
   async feasableQuote(quoteIntent: QuoteIntentModel) {
     this.logger.debug(
@@ -68,7 +56,7 @@ export class FeasableIntentService implements OnModuleInit {
     }
 
     //check if we have tokens on the solver chain
-    const { feasable, results } = await this.validateExecution(model.intent, solver)
+    const { feasable, results } = await this.feasableService.validateExecution(model.intent, solver)
     const jobId = getIntentJobId('feasable', intentHash, model!.intent.logIndex)
     this.logger.debug(
       EcoLogMessage.fromDefault({
@@ -88,174 +76,5 @@ export class FeasableIntentService implements OnModuleInit {
     } else {
       await this.utilsIntentService.updateInfeasableIntentModel(model, results)
     }
-  }
-
-  /**
-   * Validates that each target-data pair is feasible for execution.
-   *
-   * @param model the create intent model
-   * @param solver the target solver
-   * @returns
-   */
-  async validateExecution(
-    model: ValidationIntentInterface,
-    solver: Solver,
-  ): Promise<{
-    feasable: boolean
-    results: (
-      | false
-      | {
-          solvent: boolean
-          profitable: boolean
-        }
-      | undefined
-    )[]
-  }> {
-    const execs = model.route.calls.map((call) => {
-      return this.validateEachExecution(model, solver, call)
-    })
-    const results = await Promise.all(execs)
-    const feasable =
-      results.every((e) => e !== false && e !== undefined && e.solvent && e.profitable) &&
-      results.length > 0
-    return { feasable, results }
-  }
-
-  /**
-   * Validates that each target-data pair is feasible for execution. This means that
-   * the solver can execute the transaction and that transaction is profitable to the solver.
-   *
-   * @param model  the create intent model
-   * @param solver the target solver
-   * @param target the target address of the call
-   * @param data the data to send to the target
-   * @returns
-   */
-  async validateEachExecution(
-    model: ValidationIntentInterface,
-    solver: Solver,
-    call: CallDataInterface,
-  ): Promise<
-    | false
-    | {
-        solvent: boolean
-        profitable: boolean
-      }
-    | undefined
-  > {
-    const tt = this.utilsIntentService.getTransactionTargetData(model, solver, call)
-    if (tt === null) {
-      this.logger.error(
-        EcoLogMessage.withError({
-          message: `feasableIntent: Invalid transaction data`,
-          error: EcoError.FeasableIntentNoTransactionError,
-          properties: {
-            model: model,
-          },
-        }),
-      )
-      return false
-    }
-
-    switch (tt.targetConfig.contractType) {
-      case 'erc20':
-        return await this.handleErc20(tt, model, solver, call.target)
-      case 'erc721':
-      case 'erc1155':
-      default:
-        return false
-    }
-  }
-
-  /**
-   * Checks if the transaction is feasible for an erc20 token transfer.
-   *
-   * @param tt the transaction target data
-   * @param model the source intent model
-   * @param solver the target solver
-   * @param target  the target address
-   * @returns
-   */
-  async handleErc20(
-    tt: TransactionTargetData,
-    model: ValidationIntentInterface,
-    solver: Solver,
-    target: Hex,
-  ): Promise<{ solvent: boolean; profitable: boolean } | undefined> {
-    const amount = tt.decodedFunctionData.args ? (tt.decodedFunctionData.args[1] as bigint) : 0n
-    switch (tt.selector) {
-      case getERC20Selector('transfer'):
-        //check we have enough tokens to transfer on destination fullfillment
-        const balance = await this.balanceService.getTokenBalance(solver.chainID, target)
-        const solvent = balance.balance >= amount
-        //return here if we dont have enough tokens to fulfill the transfer
-        if (!solvent) {
-          return { solvent, profitable: false }
-        }
-
-        const sourceChainID = model.route.source
-        const source = this.ecoConfigService
-          .getIntentSources()
-          .find((intent) => BigInt(intent.chainID) == sourceChainID)
-        if (!source) {
-          return
-        }
-        //check that we make money on the transfer
-        const fullfillAmountUSDC = this.convertToUSDC(BigInt(solver.chainID), {
-          token: target,
-          amount,
-        })
-        const profitable = this.isProfitableErc20Transfer(
-          sourceChainID,
-          source.tokens,
-          [...model.reward.tokens],
-          fullfillAmountUSDC,
-        )
-        return { solvent, profitable }
-      default:
-        return
-    }
-  }
-
-  /**
-   * Calculates if a transfer is profitable based on the reward tokens and amounts. It converts the reward tokens to a common currency, then applies
-   * the fee to the sum of the reward tokens and amounts. If the sum is greater than the fullfill amount, then the transfer is profitable
-   *
-   * @param chainID the network to check profitability on
-   * @param acceptedTokens  the tokens that we accepte on the source
-   * @param rewardTokens  the tokens that are rewarded by the intent
-   * @param rewardAmounts  the amounts of the reward tokens
-   * @param fullfillAmountUSDC  the amount of the token to transfer on the destination chain
-   * @returns
-   */
-  isProfitableErc20Transfer(
-    chainID: bigint,
-    acceptedTokens: readonly Hex[],
-    rewardTokens: RewardTokensInterface[],
-    fullfillAmountUSDC: bigint,
-  ): boolean {
-    let sum = 0n
-    const unionTokens = rewardTokens.filter((t) => acceptedTokens.includes(t.token))
-    unionTokens.forEach((token) => {
-      sum += this.convertToUSDC(chainID, token)
-    })
-
-    //check if input tokens are acceptable and greater than + fees
-    return sum >= (fullfillAmountUSDC * this.fee) / this.FEE_BASE
-  }
-
-  /**
-   * Converts a token amount to USDC for that network, we do this in order to have a baseline
-   * to compare the profitability of the transaction
-   *
-   * TODO: right now it just returns a 1-1 conversion, we need to get the price of the token in usdc
-   *
-   * @param chainID the chain to convert the token to usdc
-   * @param token   the token to convert to usdc
-   * @returns
-   */
-  convertToUSDC(chainID: bigint, token: RewardTokensInterface): bigint {
-    //todo: get the price of the token in usdc instead of assuming 1-1 here
-    return token.amount
   }
 }
