+++ src/intent/intent.module.ts
@@ -14,6 +14,7 @@ import { MongooseModule } from '@nestjs/mongoose'
 import { SolverModule } from '../solver/solver.module'
 import { FlagsModule } from '../flags/flags.module'
 import { ValidationService } from '@/intent/validation.sevice'
+import { FeasibilityService } from '@/intent/feasibility.service'
 
 @Module({
   imports: [
@@ -29,6 +30,7 @@ import { ValidationService } from '@/intent/validation.sevice'
     CreateIntentService,
     ValidateIntentService,
     FeasableIntentService,
+    FeasibilityService,
     FulfillIntentService,
     UtilsIntentService,
     ValidationService,
